respond to tweaks proposed by heartgg, including -removal of node-fetch import -concentration of fetch functions
