+++ package.json
@@ -5,7 +5,6 @@
   "description": "A basic Plasmo extension.",
   "author": "dlaun",
   "packageManager": "npm@8.19.2",
-  "type": "module",
   "scripts": {
     "dev": "plasmo dev",
     "build": "plasmo build",
@@ -13,7 +12,6 @@
   },
   "dependencies": {
     "next": "^13.1.5",
-    "node-fetch": "^3.3.0",
     "plasmo": "^0.63.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
