--- src/data/removeLater.ts
@@ -1,14 +1,0 @@
-// Hash function to get the API key
-function unRegister(myVar: string)
-{
-    let newVar = "";
-    for (var i = 0; i < myVar.length; i++)
-    {
-        let a = myVar.charCodeAt(i);
-        a = ((a*2)-8)/2;
-        newVar = newVar.concat(String.fromCharCode(a));
-    }
-    return newVar;
-}
-
-export {unRegister};
